add light subsystem to requirements

+++ commands/kicker_shoot_when_ready.py
@@ -7,7 +7,7 @@ from subsystems.shooter import ShooterSubsystem
 class KickerShootWhenReadyCommand(commands2.Command):
     # pass in parent subsystem
     def __init__(
-        self, shooter: ShooterSubsystem, lights: pilights.PiLights, rpm: float | None
+        self, shooter: ShooterSubsystem, lights: pilights.PiLights, rpm: int | None
     ) -> None:
         super().__init__()
         self.shooter = shooter
